+++ test/test_helper.rb
@@ -1,85 +1,101 @@
-# Simplecov config has to come before literally everything else
-# Open coverage/index.html in your browser after
-# running your tests for test coverage results.
-require "simplecov"
-SimpleCov.command_name "test" + (ENV["TEST_ENV_NUMBER"] || "")
-SimpleCov.start "rails" do
-  # By default we don't include avo in coverage reports since it's not user-facing application code.
-  # If you want to get test coverage for your avo resources, you can comment out or remove the next line.
-  add_filter "/avo/"
-end
-
 ENV["RAILS_ENV"] ||= "test"
 require_relative "../config/environment"
 require "rails/test_help"
+require "capybara/rails"
+require "selenium-webdriver"
 
-# Set the default language we test in to English.
-I18n.default_locale = :en
+# Configure headless/browser mode for system tests
+HEADLESS_MODE = !ENV["OPEN_BROWSER"]
 
-# We've started loading seeds by default to try to reduce any duplication of effort trying to get the test
-# environment to look the same as the actual development and production environments. This means a consolidation
-# of setup for things like the plans available for subscriptions and which outgoing webhooks are available to users.
-require File.expand_path("../../db/seeds", __FILE__)
-
-if ENV["KNAPSACK_PRO_CI_NODE_INDEX"].present?
-  require "knapsack_pro"
-  knapsack_pro_adapter = KnapsackPro::Adapters::MinitestAdapter.bind
-  knapsack_pro_adapter.set_test_helper_path(__FILE__)
-else
-  puts "Not requiring Knapsack Pro.".yellow
-  puts "If you'd like to use Knapsack Pro make sure that you've set the environment variable KNAPSACK_PRO_CI_NODE_INDEX".yellow
-end
-
-require "sidekiq/testing"
-Sidekiq::Testing.inline!
-
-ENV["MINITEST_REPORTERS_REPORTS_DIR"] = "test/reports#{ENV["TEST_ENV_NUMBER"] || ""}"
-require "minitest/reporters"
-
-reporters = []
-
-if ENV["BT_TEST_FORMAT"]&.downcase == "dots"
-  # The classic "dot style" output:
-  # ...S..E...F...
-  reporters.push Minitest::Reporters::DefaultReporter.new
-else
-  # "Spec style" output that shows you which tests are executing as they run:
-  # UserTest
-  #   test_details_provided_should_be_true_when_details_are_provided  PASS (0.18s)
-  reporters.push Minitest::Reporters::SpecReporter.new(print_failure_summary: true)
-end
-
-# This reporter generates XML documents into test/reports that are used by CI services to tally results.
-# We add it last because doing so make the visible test output a little cleaner.
-reporters.push Minitest::Reporters::JUnitReporter.new if ENV["CI"]
-
-Minitest::Reporters.use! reporters
-
-require "parallel_tests/test/runtime_logger" if ENV["PARALLEL_TESTS_RECORD_RUNTIME"]
-
-begin
-  require "bullet_train/billing/test_support"
-  FactoryBot.definition_file_paths << BulletTrain::Billing::TestSupport::FACTORY_PATH
-  FactoryBot.reload
-rescue LoadError
-end
-
-begin
-  require "bullet_train/billing/stripe/test_support"
-  FactoryBot.definition_file_paths << BulletTrain::Billing::Stripe::TestSupport::FACTORY_PATH
-  FactoryBot.reload
-rescue LoadError
-end
-
-ActiveSupport::TestCase.class_eval do
+class ActiveSupport::TestCase
   # Run tests in parallel with specified workers
-  # parallelize(workers: :number_of_processors)
+  parallelize(workers: :number_of_processors)
 
+  # Setup all fixtures in test/fixtures/*.yml
   fixtures :all
 
   # Add more helper methods to be used by all tests here...
+  include FactoryBot::Syntax::Methods
 end
 
-class ActiveSupport::TestCase
-  include FactoryBot::Syntax::Methods
+class ActionDispatch::SystemTestCase
+  # Make the Capybara DSL available in system tests
+  include Capybara::DSL
+  include Warden::Test::Helpers
+
+  # Helper methods for training program tests
+  def complete_video_module
+    # Wait for video to load and simulate completion
+    find("video")
+    page.execute_script("document.querySelector('video').dispatchEvent(new Event('ended'))")
+    wait_for_ajax
+  end
+
+  def complete_quiz(correct: true)
+    within(".question-list") do
+      all(".answer-option").each do |option|
+        if correct && option.text.include?("Correct Answer")
+          option.click
+        elsif !correct && option.text.include?("Wrong Answer")
+          option.click
+        end
+      end
+      click_button "Submit Answers"
+    end
+    wait_for_ajax
+  end
+
+  def wait_for_ajax
+    Timeout.timeout(Capybara.default_max_wait_time) do
+      loop until finished_all_ajax_requests?
+    end
+  end
+
+  def finished_all_ajax_requests?
+    page.evaluate_script("jQuery.active").zero?
+  end
+
+  def toggle_browser_mode(open_browser)
+    ENV["OPEN_BROWSER"] = open_browser ? "1" : nil
+    Capybara.current_driver = Capybara.javascript_driver
+  end
+
+  # Clean up any resources after each test
+  def teardown
+    super
+    Capybara.reset_sessions!
+    Capybara.use_default_driver
+  end
+end
+
+# Configure VCR for recording HTTP interactions
+require "vcr"
+VCR.configure do |config|
+  config.cassette_library_dir = "test/vcr_cassettes"
+  config.hook_into :webmock
+  config.ignore_localhost = true
+  config.allow_http_connections_when_no_cassette = false
+end
+
+# Configure test coverage reporting
+require "simplecov"
+SimpleCov.start "rails" do
+  add_filter "/test/"
+  add_filter "/config/"
+
+  add_group "Controllers", "app/controllers"
+  add_group "Models", "app/models"
+  add_group "Components", "app/javascript/components"
+  add_group "Stores", "app/javascript/stores"
+end
+
+# Load support files
+Dir[Rails.root.join("test/support/**/*.rb")].each { |f| require f }
+
+# Configure time helpers
+include ActiveSupport::Testing::TimeHelpers
+
+# Configure test screenshots
+if ENV["SAVE_SCREENSHOTS"]
+  Capybara.save_path = Rails.root.join("tmp/screenshots")
 end